requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
python-engineio==4.8.0
python-socketio==5.10.0
qrcode==7.4.2
Pillow>=10.2.0
Flask-CORS==4.0.0
gunicorn==20.1.0
